#47 Feat: ProductCard component
