+++ practice-two/src/components/common/material/Card.jsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+// Card component
+const Card = ({
+  maxW = 'xl',
+  variant = 'primary',
+  children,
+  onChange = () => {},
+}) => {
+  const commonClasses = 'bg-white';
+
+  const variantClasses = {
+    primary: 'border-none',
+    secondary: 'border-secondary',
+    success: 'border-green-500',
+    danger: 'border-red-500',
+    warning: 'border-yellow-500',
+  };
+
+  const cardClasses = `${commonClasses} ${variantClasses[variant]}`;
+
+  return (
+    <div className={`card max-w-${maxW} ${cardClasses}`} onChange={onChange}>
+      {children}
+    </div>
+  );
+};
+
+Card.propTypes = {
+  maxW: PropTypes.string,
+  variant: PropTypes.oneOf([
+    'primary',
+    'secondary',
+    'success',
+    'danger',
+    'warning',
+  ]),
+  onChange: PropTypes.func,
+  children: PropTypes.node.isRequired,
+};
+
+export default Card;
