+++ practice-two/src/components/common/material/CardOverflow.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+// CardOverflow component
+const CardOverflow = ({ children }) => (
+  <div className="card-overflow">{children}</div>
+);
+
+CardOverflow.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default CardOverflow;
